+++ Changes.md
@@ -4,6 +4,8 @@
 
 - Supports active state for night vision control
 - Supports alternative (optional) configuration when night vision on
+- Supports all combinations of two colours for flashing light modes in
+  configuration
 - Make alternative hardpoints deployed configuration optional in the
   configuration file; falls back to default configuration from the file if
   missing